remove kludge, collapse everything to indexer.js middleware
