+++ middleware/indexer.js
@@ -1,4 +1,3 @@
-const HELPER_DISABLE_INDEXER = process.env.HELPER_DISABLE_INDEXER === 'yes' || process.env.HELPER_DISABLE_INDEXER === 'true';
 const { Client } = require('pg');
 
 let client;
@@ -13,7 +12,29 @@ async function getPgClient() {
     return client;
 }
 
-async function findAccountsByPublicKeyIndexer(ctx) {
+async function getStakingTxs(ctx) {
+    const { accountId } = ctx.params;
+
+    const client = await getPgClient();
+    const { rows } = await client.query(`
+        select 
+        receiver_account_id validator_id,
+        args->>'method_name' method_name,
+        coalesce(nullif(args->>'deposit', '0'), convert_from(decode(args->>'args_base64', 'base64'), 'utf8')::jsonb->>'amount', '0') amount,
+        included_in_block_timestamp ts
+        from receipts
+        inner join action_receipt_actions using(receipt_id)
+        where
+        predecessor_account_id = $1 and
+        action_kind = 'FUNCTION_CALL' and 
+        args->>'method_name' in ('stake', 'deposit_and_stake', 'unstake', 'unstake_all')
+        limit 9999
+    `, [accountId]);
+    
+    ctx.body = rows;
+}
+
+async function findAccountsByPublicKey(ctx) {
     const { publicKey } = ctx.params;
 
     const client = await getPgClient();
@@ -21,16 +42,7 @@ async function findAccountsByPublicKeyIndexer(ctx) {
     ctx.body = rows.map(({ account_id }) => account_id);
 }
 
-// TODO: Remove the kludge when indexer is working well
-const models = require('../models');
-async function findAccountsByPublicKeyTemp(ctx) {
-    const { publicKey } = ctx.params;
-
-    const rows = await models.AccountByPublicKey.findAll({ where: { publicKey }, attributes: ['accountId'], group: ['accountId'] });
-    ctx.body = rows.map(({ accountId }) => accountId);
-}
-
 module.exports = {
-    findAccountsByPublicKey: HELPER_DISABLE_INDEXER ? findAccountsByPublicKeyTemp : findAccountsByPublicKeyIndexer,
-    findAccountsByPublicKeyIndexer
+    findAccountsByPublicKey,
+    getStakingTxs
 };
